EEP2/IEP-parse/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pdfminer.six==20221105
python-docx==0.8.11
python-pptx==0.6.21
gunicorn==20.1.0
PyMuPDF==1.23.3
